Tests for spec harness passing

+++ db/migrate/20151103041152_create_items.rb
@@ -4,7 +4,7 @@ class CreateItems < ActiveRecord::Migration
     create_table :items do |t|
       t.citext :name
       t.text :description
-      t.integer :unit_price
+      t.decimal :unit_price
       t.references :merchant, index: true, foreign_key: true
 
       t.timestamps null: false
